Fixed bug that caused xml:lang to be re-defined in child nodes.
The recent changes created a bug that caused XMLLiterals to re-define
xml:lang when an XMLLiteral contained child elements. Fixed so that
the context carries a flag that notifies the XMLLiteral code whether
or not an xml:lang was inserted previously when building the XMLLiteral.
This still allows the XMLLiteral to override the inserted xml:lang value.