Fix buffer sizing and terminate with NUL always.
(rdfa_init_base): Add +1 to realloc and NUL terminating buffer so
that strstr() can legally work, it expects to find a NUL.

(rdfa_parse_start): Use malloc for buffer malloc - save lots of
zeroing that isn't used.  +1 to length for a NUL and set it. Do
not need to multiply size by sizeof(char) again, it's already
calculated in context->wb_allocated.
This fixes Ticket#37
http://rdfa.digitalbazaar.com/librdfa/trac/ticket/37