Ensure reallocated buffer size is big enough
(rdfa_init_base): If the bytes_read is bigger than READ_BUFFER_SIZE,
when the buffer size is increased, it must be made sufficiently
large.

Previously the logic created a buffer that was too small so if
bytes_read was > 4K, it would memmove() into memory it wasn't allowed
to.  (Also the NUL termination was wrong, fixed in a previous commit).